ReAct/requirements.txt
openai==0.28.1
requests==2.31.0
python-dotenv==1.0.0
